+++ src/main/java/seedu/chatbaby/AddCommand.java
@@ -1,4 +1,17 @@
-package com.chatbaby;
+package chatbaby;
+
+import chatbaby.Command;
+import chatbaby.Storage;
+import chatbaby.Task;
+import chatbaby.TaskType;
+import chatbaby.TaskList;
+import chatbaby.Ui;
+import chatbaby.Parser;
+import chatbaby.ChatBabyException;
+import chatbaby.Deadline;
+import chatbaby.Event;
+import chatbaby.ToDo;
+
 
 import java.time.format.DateTimeParseException;
 import java.util.ArrayList;
