+++ src/main/java/seedu/chatbaby/ChatBabyException.java
@@ -1,4 +1,4 @@
-package com.chatbaby;
+package chatbaby;
 
 public class ChatBabyException extends Exception{
     public ChatBabyException(String message) {
