+++ src/main/java/seedu/chatbaby/Command.java
@@ -1,4 +1,16 @@
-package com.chatbaby;
+package chatbaby;
+
+import chatbaby.Command;
+import chatbaby.Storage;
+import chatbaby.Task;
+import chatbaby.TaskType;
+import chatbaby.TaskList;
+import chatbaby.Ui;
+import chatbaby.Parser;
+import chatbaby.ChatBabyException;
+import chatbaby.Deadline;
+import chatbaby.Event;
+import chatbaby.ToDo;
 
 public abstract class Command {
     String commandBody;
