+++ src/main/java/seedu/chatbaby/ExitCommand.java
@@ -1,5 +1,16 @@
-package com.chatbaby;
+package chatbaby;
 
+import chatbaby.Command;
+import chatbaby.Storage;
+import chatbaby.Task;
+import chatbaby.TaskType;
+import chatbaby.TaskList;
+import chatbaby.Ui;
+import chatbaby.Parser;
+import chatbaby.ChatBabyException;
+import chatbaby.Deadline;
+import chatbaby.Event;
+import chatbaby.ToDo;
 public class ExitCommand extends Command {
     public ExitCommand(String commandBody) {
         super(commandBody);
