+++ src/main/java/seedu/chatbaby/ListCommand.java
@@ -1,0 +1,22 @@
+package chatbaby;
+
+import chatbaby.Command;
+import chatbaby.Storage;
+import chatbaby.Task;
+import chatbaby.TaskType;
+import chatbaby.TaskList;
+import chatbaby.Ui;
+import chatbaby.Parser;
+import chatbaby.ChatBabyException;
+import chatbaby.Deadline;
+import chatbaby.Event;
+import chatbaby.ToDo;
+public class ListCommand extends Command {
+    public ListCommand(String commandBody) {
+        super(commandBody);
+    }
+    @Override
+    public void execute(TaskList tasks, Ui ui, Storage storage) {
+        tasks.listTasks();
+    }
+}
