+++ src/main/java/seedu/chatbaby/Storage.java
@@ -1,5 +1,16 @@
-package com.chatbaby;
+package chatbaby;
 
+import chatbaby.Command;
+import chatbaby.Storage;
+import chatbaby.Task;
+import chatbaby.TaskType;
+import chatbaby.TaskList;
+import chatbaby.Ui;
+import chatbaby.Parser;
+import chatbaby.ChatBabyException;
+import chatbaby.Deadline;
+import chatbaby.Event;
+import chatbaby.ToDo;
 import java.io.*;
 import java.util.ArrayList;
 import java.util.Scanner;
