+++ src/main/java/seedu/chatbaby/TaskType.java
@@ -1,4 +1,4 @@
-package com.chatbaby;
+package chatbaby;
 
 public enum TaskType {
     TODO, DEADLINE, EVENT;
